Private runners creation support, tooling setup improvement and TF pipeline simplifications, removal of PAT requirements (#299)
This PR does the following things:

- Adds support for the creation and configuration of private runners for
both AzDO and GitHub workflows.
- Removes the need for a Github PAT or an AzDO PAT, during pipeline
config, as long as the `gh` cli or `az` cli are signed in.
- Fixes several issues during the tooling setup on brand new private
runners
- Moves to Terraform 1.11.0 as the required version, since it has much
better support for azurerm backend using az cli credentials. Leveraging
this simplifies a lot the authentication process for terraform
pipelines, avoiding the need for specific auth code or env variables.

Fixes #298, #274

---------

Co-authored-by: Pablo Zaidenvoren <pablozaiden@users.noreply.github.com>
Co-authored-by: Hattan Shobokshi <hattan.shobokshi@microsoft.com>